enforce ctg_hex_id.astype(int) before idxmax
Otherwise, ctg_hex_id may not be int sometimes and idxmax would raise error

+++ kleat/post.py
@@ -95,7 +95,7 @@ def agg_polya_evidence_per(grp):
     any_cols = grp[S.COLS_TO_ANY].any()
     str_cols = grp[S.COLS_TO_JOIN].apply(set_sort_join_strs)
     # pick the strongest PAS hexamer
-    hex_cols = grp[S.COLS_CONTIG_HEXAMERS].loc[grp.ctg_hex_id.idxmax()]
+    hex_cols = grp[S.COLS_CONTIG_HEXAMERS].loc[grp.ctg_hex_id.astype(int).idxmax()]
     one_cols = grp[S.COLS_PICK_ONE].iloc[0]
     return pd.concat([sum_cols, max_cols, any_cols,
                       str_cols, hex_cols, one_cols])
